Add files via upload
Now this is the programming of the app itself, programmed using Python and MySQL connector. This is the implementation of the whole Video-Store-Application creation. It is very important to read the User Documentation word document to learn how to install and setup the application as well as the whole description of the App.